chore(git-pull-service): refactor prisma client folder structure
