+++ ee/packages/amplication-git-pull-service/.gitignore
@@ -1,1 +1,1 @@
-src/generated-prisma-client
+generated-prisma-client
