+++ ee/packages/amplication-git-pull-service/prisma/schema.prisma
@@ -1,6 +1,6 @@
 generator client {
   provider = "prisma-client-js"
-  output   = "../src/generated-prisma-client"
+  output   = "./generated-prisma-client"
 }
 
 datasource db {
