+++ ee/packages/amplication-git-pull-service/project.json
@@ -59,7 +59,7 @@
     },
     "db:prisma:generate": {
       "executor": "nx:run-commands",
-      "outputs": ["src/generated-prisma-client"],
+      "outputs": ["prisma/generated-prisma-client"],
       "options": {
         "command": "prisma generate --schema ./prisma/schema.prisma",
         "cwd": "ee/packages/amplication-git-pull-service"
