+++ ee/packages/amplication-git-pull-service/scripts/customSeed.ts
@@ -1,4 +1,4 @@
-import { PrismaClient } from "../src/generated-prisma-client";
+import { PrismaClient } from "../prisma/generated-prisma-client";
 
 export async function customSeed() {
   const client = new PrismaClient();
