+++ ee/packages/amplication-git-pull-service/src/prisma/prisma.service.ts
@@ -1,5 +1,5 @@
 import { INestApplication, Injectable, OnModuleInit } from "@nestjs/common";
-import { PrismaClient } from "../generated-prisma-client";
+import { PrismaClient } from "../../prisma/generated-prisma-client";
 
 @Injectable()
 export class PrismaService extends PrismaClient implements OnModuleInit {
